+++ src/FormattingSettingsService.ts
@@ -29,7 +29,7 @@ export class FormattingSettingsService implements IFormattingSettingsService {
                 if (card instanceof CompositeCard) card.topLevelSlice?.setPropertiesValues(dataViewObjects, card.name);
 
                 const cardGroupInstances = <CardGroupEntity[]>(card instanceof SimpleCard ? [ card ] : card.groups);
-                cardGroupInstances.forEach((cardGroupInstance: CardGroupEntity) => {
+                cardGroupInstances?.forEach((cardGroupInstance: CardGroupEntity) => {
                     // Set current top level toggle value
                     cardGroupInstance.topLevelSlice?.setPropertiesValues(dataViewObjects, card.name);
                     cardGroupInstance?.slices?.forEach((slice: Slice) => {
@@ -83,8 +83,7 @@ export class FormattingSettingsService implements IFormattingSettingsService {
                 const cardGroupInstances = <CardGroupEntity[]>(isSimpleCard ? 
                     [ card ].filter(({visible = true}) => visible) : 
                     card.groups.filter(({visible = true}) => visible));
-                cardGroupInstances
-                    .forEach((cardGroupInstance: CardGroupEntity) => {
+                cardGroupInstances?.forEach((cardGroupInstance: CardGroupEntity) => {
                         const groupUid = cardGroupInstance.name + "-group";
 
                         // Build formatting group for each group
@@ -121,6 +120,9 @@ export class FormattingSettingsService implements IFormattingSettingsService {
                             }
 
                             container.containerItems.forEach((containerItem: formattingSettings.ContainerItem) => {
+                                if(!containerItem) { // This is to prevent error when container item is null or undefined
+                                    return;
+                                }
                                 // Build formatting container item object
                                 const containerIemName = containerItem.displayNameKey ? containerItem.displayNameKey : containerItem.displayName;
                                 const containerItemUid: string = containerUid + containerIemName;
@@ -193,7 +195,7 @@ export class FormattingSettingsService implements IFormattingSettingsService {
         const cardGroupInstances = <CardGroupEntity[]>(card instanceof SimpleCard ? 
             [ card ].filter(({visible = true}) => visible) : 
             card.groups.filter(({visible = true}) => visible));
-        cardGroupInstances.forEach((cardGroupInstance: CardGroupEntity) => {
+        cardGroupInstances?.forEach((cardGroupInstance: CardGroupEntity) => {
             cardSlicesDefaultDescriptors = this.getSlicesRevertToDefaultDescriptor(card.name, cardGroupInstance.slices, sliceNames, cardGroupInstance.topLevelSlice);
 
             cardGroupInstance.container?.containerItems?.forEach((containerItem: formattingSettings.ContainerItem) => {